WASM.mjs template fix for Node/Browser detection
fetch === 'undefined' doesn't work for my setup when running ava in the ICU4X ffi/capi folder. fetch is defined for the Browser as well as Node.JS.

typeof process hopefully works better to detect if it's Node.JS instead of Browser modules? But let me know if this breaks anything for anyone else!