Add support for DiplomatOption
Because we currently don't pass context when we recursively evaluate parameters, DiplomatOption effectively is ignored for the purposes of demonstration.

Especially difficult to resolve since we flatten all struct and opaque construction into one render terminus (making it hard to set up optional support), so this is a workaround for now.