Nanobind reference leak fix (#912)
Not sure how `nb::keep_alive` works, but the current way it's being used
in the nanobind bindings is leading to reference leaks for returning
references.

This fixes that (and adds a failing test to check nanobind for memory
leaks for some functions), but @Walter-Reactor your thoughts here are
appreciated.

---------

Co-authored-by: Walter Gray <walter@0m.dev>