&[Struct] support (#908)
Fix for https://github.com/rust-diplomat/diplomat/issues/769, adds
support for `&[Struct]` types. With the limitation that the Structs must
only contain non-slice primitives (currently need to actually put checks
for this limitation in HIR code).

Currently this is just the rough outline of the setup inside the
AST/HIR. A lot of how this gets implemented will be down to the
backends. Let me know if this matches roughly the skeleton proposed with
the original issue.

I plan on also adding a sample implementation for C++/Nanobind.

---------

Co-authored-by: Walter Gray <walter@0m.dev>