Nanobind: Keep Owned Non-Self References Alive (#925)
Currently this is just a failing test, I'm still working out what the
fix needs to be.

You can fool the garbage collector in Python into collecting data that's
still being used elsewehere (like say, a reference to an inner vec).
This can be fixed with a `nb::keep_alive<0, 1>()` on the returned
reference to the inner data, but I still need to research what scenarios
we want to implement `keep_alive` for.