Nanobind Overloading Support (#928)
Since we never know when an overload will happen (and all overloaded
functions need to know their argument types), this requires us to call
`gen_type_name` consistently for all functions, which may destroy some
of the functionality Nanobind is trying to avoid.

I can either leave this as-is, or maybe we could re-factor how
`includes` are used within `gen_type_name`, maybe added to `ParamInfo`?
Then the includes are only processed if we know for certain they will be
used in the function.