Improved Macro arg parsing (#923)
This greatly expands `macros.rs` to allow for more complex parsing of
arguments. A lot of this is built on the assumption that rustc will
catch the superset of all errors related to macro parsing, so we don't
have to match that superset. We're also heavily reliant on `syn` to
parse a lot of this Rust syntax for us.

This still does not support multiple arms, or `$($...)*?+` parsing. But
it's a decent step towards that sort of support.

TODOs:

- [x] Rigorous testing of each `MacroFragSpec` type.
- [X] ~~`pat` MacroFragSpec parsing~~ (Parsing looks to be too
complicated for right now, can be done in a subsequent PR)
- [X] MacroMatcher sub-parsing (basically more delimited parsing)
- [ ] File issues for follow ups