Fix bug with handling optionals in JS (#953)
Fixes https://github.com/unicode-org/icu4x/issues/6688

The core problem is the distinction between TypedArray and ArrayBuffer:
TypedArrays are views into an ArrayBuffer, whereas an ArrayBuffer is
often a very large list of bytes. When working with wasm, we typically
work with the underlying ArrayBuffer and use offset "pointers".

The source of the bug was two separate bugs cancelling each other out.

Firstly, we were accidentally giving Rust a pointer value 0 because we
forgot to return anything from `optionToBufferForCalling()` and were
passing it `undefined`. Rust was reading from the beginning of Wasm
memory.

Secondly, we were accidentally writing data to pointer value 0 because
we were passing down the ArrayBuffer with offset 0 to the callback,
thinking we had already offset it when making a TypedArray, except
`TypedArray.buffer` returns the original full buffer, not some slice
view. So we were writing to the beginning of Wasm memory as well.
Furthermore, when we made the TypedArray, we passed it `buf` as the
offset which also probably turned into 0, so the Option discriminant got
written to the beginning of wasm memory (but offset a bit).

These cancel out so it worked as long as we only passed one optional
parameter, using the beginning of wasm memory as "scratch space".

We should probably be more consistent about "buffer" terminology in this
code.