Add RustWriteVec, a safe Rust DiplomatWrite wrapper (#898)
Fixes #819

After a lot of thought, I made a new type, `DiplomatWriteVec`, instead
of using `DiplomatWrite`, because:

1. This allows a cleaner abstraction over the C functions, which is what
we generally want to test
2. It's not clear if we'd want a `DiplomatWrite` or a
`Box<DiplomatWrite>`; with the wrapper type, we have flexibility to
change this
3. I found it easier to reason about the safety as a client of
`DiplomatWrite` instead of adding too many functions to `DiplomatWrite`
itself
4. The Drop impl depends on whether the DiplomatWrite was initialized
with a Vec or not. Using a separate type lets me guarantee that the
DiplomatWrite was created in that way.

The new type is returned by `DiplomatWrite::new_vec()` so most users
won't necessarily see it.

For now, it only exposes `.as_bytes()`, but in the future we could
possibly add `.into_vec()` or similar. We may be able to add `.as_str()`
but I didn't want to think too hard about that invariant right now;
there were enough other ones to handle.

I took time to go through the safety checklists, since there is a lot of
unsafe code here, but please double-check my work.

`diplomat-runtime` is still in a pre-1.0 state, so I think this is
landable even if we want to continue to debate the API shape, so long as
the code is sound.

---------

Co-authored-by: Manish Goregaokar <manishsmail@gmail.com>