Add stable ABI C++ type for string view in slice (#964)
Fixes #963 

This adds `diplomat::string_view_for_slice` to the c++ runtime which
internally contains `DiplomatStringView` so has the right layout to pass
over the ABI in a slice, at the downside of needing to use this type
instead of `std::string_view` for this case.

This also adds an appropriate adapter to nanobind and updates the tests.

---------

Co-authored-by: Manish Goregaokar <manishsmail@gmail.com>