Only use wasm_glue on wasm32-unknown-unknown, not WASI (#977)
The functional change here is a one-line fix, with which all
functionality (at least as covered by tests) works under both WASIp1 and
p2.

The second commit contains changes to add CI coverage. I'm happy to
remove that, or only test one of the targets, if you don't want to spend
the CI resources, of course.