Upgrade dart (#982)
The changes are mainly due to splitting of `native_assets_cli` into
`hooks` and `code_assets`, and the enabling of the native assets
experiment by default on `beta`. If we want to use a stable version,
that's also possible, but then we still need to pass the
`enable-experiment` flag for native assets.