Add Maven feature tests (#994)
Useful for people trying to get stuff working with Maven


(Also, I am still unable to get the Gradle build working locally)

Would be cool to try and get
https://github.com/questdb/rust-maven-plugin working.

Note that this only adds it for feature_tests, not for example. I think
that's fine for now, I didn't want to add more sets of CI tasks.