Kotlin fixes around Option and Result returns (#995)
There were three bugs in the Kotlin code:

- Enums have an underlying FFI type of `Int`, and we were initializing
them incorrectly as `Int()` in some places since we were treating them
the same as struct
- Because they have the same underlying FFI type, we might end up with
multiple Result types with the same name, but different defaults.
Tweaked them to generate. A better design would be to store a map of
Result _native_ types to their potential defaults, but that's too big of
a refactor for now (cc @jcrist1 if you're interested)
- Option returns were just doing the wrong thing and not returning
anything.